dz2/requirements.txt
pyyaml==6.0
GitPython==3.1.31
graphviz==0.20.1
pytest==7.4.0
